requirements.txt
openai>=1.40.0

langfuse>=3.0.0,<4.0.0

python-dotenv>=1.0.1

pandas>=2.1.0

numpy>=1.26.0

matplotlib>=3.8.0

Pillow>=10.0.0

opencv-python>=4.9.0.80

streamlit>=1.35.0

plotly>=5.18.0

tqdm>=4.66.0

pytest>=8.0.0
